Run an object

Run an object

run(x, ...)

# S3 method for BenchmarkDataFrame
run(x, ..., publish = FALSE, run_name = NULL, run_reason = NULL)

Arguments

x

An S3 classed object to run

...

Additional arguments passed through to methods. For run.BenchmarkDataFrame, passed through to get_default_parameters() (when parameters are not specified) and run_benchmark().

publish

Flag for whether to publish results to a Conbench server. See "Environment Variables" section for how to specify server details. Requires the benchconnect CLI is installed; see install_benchconnect().

run_name

Name for the run. If not specified, will use {run_reason}: {commit hash}

run_reason

Required. Low-cardinality reason for the run, e.g. "commit" or "test"

Value

A modified object containing run results. For run.BenchmarkDataFrame, a results list column is appended.

Environment Variables

CONBENCH_URL: Required. The URL of the Conbench server with no trailing slash. For arrow, should be https://conbench.ursa.dev.

CONBENCH_EMAIL: The email to use for Conbench login. Only required if the server is private.

CONBENCH_PASSWORD: The password to use for Conbench login. Only required if the server is private.

CONBENCH_PROJECT_REPOSITORY: The repository name (in the format org/repo) or the URL (in the format https://github.com/org/repo). Defaults to "https://github.com/apache/arrow" if unset.

CONBENCH_PROJECT_PR_NUMBER: Recommended. The number of the GitHub pull request that is running this benchmark, or NULL if it's a run on the default branch

CONBENCH_PROJECT_COMMIT: The 40-character commit SHA of the repo being benchmarked. If missing, will attempt to obtain it from arrow::arrow_info()$build_info$git_id, though this may not be populated depending on how Arrow was built.

CONBENCH_MACHINE_INFO_NAME: Will override detected machine host name sent in machine_info.name when posting runs and results. Needed for cases where the actual host name can vary, like CI and cloud runners.